Add sass support and add template

+++ build.sh
@@ -1,11 +1,22 @@
 #!/bin/bash
 
-printf "Cleaning build directory ..."
-rm -r build/*
-printf "done!\n"
+if [ -d ./build ]; then
+  printf "Cleaning build directory ..."
+  rm -r build/*
+  printf "done!\n"
+else
+  printf "Creating build directory..."
+  mkdir ./build
+  printf "done!\n"
+fi
 
 printf "Compiling coffee..."
-coffee --output build -c src
+coffee -c --output build src
+printf "done!\n"
+
+printf "Compiling sass..."
+mkdir build/stylesheets
+sass --trace src/stylesheets/base.sass build/stylesheets/base.css
 printf "done!\n"
 
 printf "Generating HAML..."
